Edited such that magnetometer readings can be taken!
Previously on the serial monitor magnetometer readings were showing up as 0.0, 0.0, 0.0. Problem was due to a float precision issue

+++ Guidance, Navigation and Control/Flight Code/include/constants.h
@@ -61,16 +61,16 @@ struct FlightData {
     float rotZ;            
 
     /** @brief [\muT] Strength of magnetic field on the rocket's X-Axis */
-    float magX;
+    double magX;
 
     /** @brief [\muT] Strength of magnetic field on the rocket's Y-Axis */
-    float magY;
+    double magY;
 
     /** @brief [\muT] Strength of magnetic field on the rocket's Z-Axis */
-    float magZ;
+    double magZ;
 
     /** @brief [rad] Magnetic heading relative to true north. */
-    float heading;
+    double heading;
     
     /** @brief [m/s^2] Total magnitude of the acceleration vector (G-force). */
     float accelMagnitude;
